// frontend/src/models/collateralStatus.ts
export enum CollateralStatus {
  PENDING = "pending",
  PENDIENTE = "Pendiente",
  APPROVED = "approved",
  APROBADO = "Aprobado",
  REJECTED = "rejected",
  RECHAZADO = "Rechazado",
  UNDER_REVIEW = "under review",
  ACTIVE = "active",
  ACTIVO = "Activo",
  INACTIVE = "inactive",
  INACTIVO = "Inactivo",
  SEIZED = "seized",
  INCAUTADO = "Incautado",
  RELEASED = "released",
  LIBERADO = "Liberado",
  DEFAULTED = "defaulted",
  INCUMPLIDO = "Incumplido",
  CLEARED = "cleared",
  LIQUIDADO = "Liquidado",
  EXPIRED = "expired",
  VENCIDO = "Vencido",
  ON_HOLD = "on hold",
  EN_ESPERA = "En espera",
}